+++ sdk/keyvault/azure-security-keyvault-secrets/src/main/java/com/azure/security/keyvault/secrets/implementation/KeyVaultCredentialPolicy.java
@@ -12,6 +12,8 @@ import com.azure.core.util.CoreUtils;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
+import java.net.URI;
+import java.net.URISyntaxException;
 import java.net.URL;
 import java.nio.ByteBuffer;
 import java.util.Collections;
@@ -34,8 +36,8 @@ public class KeyVaultCredentialPolicy extends BearerTokenAuthenticationPolicy {
     private static final String KEY_VAULT_STASHED_CONTENT_KEY = "KeyVaultCredentialPolicyStashedBody";
     private static final String KEY_VAULT_STASHED_CONTENT_LENGTH_KEY = "KeyVaultCredentialPolicyStashedContentLength";
     private static final String WWW_AUTHENTICATE = "WWW-Authenticate";
-    private static final ConcurrentMap<String, String> SCOPE_CACHE = new ConcurrentHashMap<>();
-    private String scope;
+    private static final ConcurrentMap<String, ChallengeParameters> CHALLENGE_CACHE = new ConcurrentHashMap<>();
+    private ChallengeParameters challenge;
 
     /**
      * Creates a {@link KeyVaultCredentialPolicy}.
@@ -80,6 +82,7 @@ public class KeyVaultCredentialPolicy extends BearerTokenAuthenticationPolicy {
      *
      * @param authenticateHeader The authentication header containing all the challenges.
      * @param authChallengePrefix The authentication challenge name.
+     *
      * @return A boolean indicating if the challenge is a bearer challenge or not.
      */
     private static boolean isBearerChallenge(String authenticateHeader, String authChallengePrefix) {
@@ -92,15 +95,17 @@ public class KeyVaultCredentialPolicy extends BearerTokenAuthenticationPolicy {
         return Mono.defer(() -> {
             HttpRequest request = context.getHttpRequest();
 
-            // If this policy doesn't have an authorityScope cached try to get it from the static challenge cache.
-            if (this.scope == null) {
+            // If this policy doesn't have challenge parameters cached try to get it from the static challenge cache.
+            if (this.challenge == null) {
                 String authority = getRequestAuthority(request);
-                this.scope = SCOPE_CACHE.get(authority);
+                this.challenge = CHALLENGE_CACHE.get(authority);
             }
 
-            if (this.scope != null) {
-                // We fetched the scope from the cache, but we have not initialized the scopes in the base yet.
-                TokenRequestContext tokenRequestContext = new TokenRequestContext().addScopes(this.scope);
+            if (this.challenge != null) {
+                // We fetched the challenge from the cache, but we have not initialized the scopes in the base yet.
+                TokenRequestContext tokenRequestContext = new TokenRequestContext()
+                    .addScopes(this.challenge.getScopes())
+                    .setTenantId(this.challenge.getTenantId());
 
                 return setAuthorizationHeader(context, tokenRequestContext);
             }
@@ -150,33 +155,92 @@ public class KeyVaultCredentialPolicy extends BearerTokenAuthenticationPolicy {
             }
 
             if (scope == null) {
-                this.scope = SCOPE_CACHE.get(authority);
+                this.challenge = CHALLENGE_CACHE.get(authority);
 
-                if (this.scope == null) {
+                if (this.challenge == null) {
                     return Mono.just(false);
                 }
             } else {
-                this.scope = scope;
+                String authorization = challengeAttributes.get("authorization");
 
-                SCOPE_CACHE.put(authority, this.scope);
+                if (authorization == null) {
+                    authorization = challengeAttributes.get("authorization_uri");
+                }
+
+                final URI authorizationUri;
+
+                try {
+                    authorizationUri = new URI(authorization);
+                } catch (URISyntaxException e) {
+                    // The challenge authorization URI is invalid.
+                    return Mono.just(false);
+                }
+
+                this.challenge = new ChallengeParameters(authorizationUri, new String[] { scope });
+
+                CHALLENGE_CACHE.put(authority, this.challenge);
             }
 
-            TokenRequestContext tokenRequestContext = new TokenRequestContext().addScopes(this.scope);
+            TokenRequestContext tokenRequestContext = new TokenRequestContext()
+                .addScopes(this.challenge.getScopes())
+                .setTenantId(this.challenge.getTenantId());
 
             return setAuthorizationHeader(context, tokenRequestContext)
                 .then(Mono.just(true));
         });
     }
 
-    static void clearCache() {
-        SCOPE_CACHE.clear();
+    private static class ChallengeParameters {
+        private final URI authorizationUri;
+        private final String tenantId;
+        private final String[] scopes;
+
+        ChallengeParameters(URI authorizationUri, String[] scopes) {
+            this.authorizationUri = authorizationUri;
+            tenantId = authorizationUri.getPath().split("/")[1];
+            this.scopes = scopes;
+        }
+
+        /**
+         * Get the {@code authorization} or {@code authorization_uri} parameter from the challenge response.
+         */
+        public URI getAuthorizationUri() {
+            return authorizationUri;
+        }
+
+        /**
+         * Get the {@code resource} or {@code scope} parameter from the challenge response. This should end with
+         * "/.default".
+         */
+        public String[] getScopes() {
+            return scopes;
+        }
+
+        /**
+         * Get the tenant ID from {@code authorizationUri}.
+         */
+        public String getTenantId() {
+            return tenantId;
+        }
     }
 
+    public static void clearCache() {
+        CHALLENGE_CACHE.clear();
+    }
+
+    /**
+     * Gets the host name and port of the Key Vault or Managed HSM endpoint.
+     *
+     * @param request The {@link HttpRequest} to extract the host name and port from.
+     *
+     * @return The host name and port of the Key Vault or Managed HSM endpoint.
+     */
     private static String getRequestAuthority(HttpRequest request) {
         URL url = request.getUrl();
         String authority = url.getAuthority();
         int port = url.getPort();
 
+        // Append port for complete authority.
         if (!authority.contains(":") && port > 0) {
             authority = authority + ":" + port;
         }